+++ .env
@@ -1,2 +1,2 @@
-API_KEY="uGX4r9523386"
+API_KEY="Z9ji76HN3387"
 URL="https://alphagov.github.io/accessibility-tool-audit/test-cases.html"
